+++ admin/footer.php
@@ -1,0 +1,6 @@
+<script src="js/jquery-3.3.1.min.js"></script>
+<script src="js/jquery-ui-1.12.1.min.js"></script>
+<script src="js/bootstrap-4.1.0.min.js"></script>
+<script src="js/yap.js"></script>
+</body>
+</html>